Bugfix: Allow viewing logs from login screen
In legacy portions of the code, `core.connection.isOpen` is used as a proxy for
having loaded `core.state.settings`, and it works only because you're stuck on
a load screen for the duration of both structures setting up. However, this
coupling is just tip-toeing around the issue. `.settings` DELIBERATELY throws,
so it's more convergent to catch than to pretend we care about `.isOpen`.