Utilize a different technique for opening web browser
This was a security concern in the original f-chat but wasn't a concern for Frolic due to a redesign of this system many months ago. Instead, this touch up is being done as a result of research prompted by an abundance of caution. Turns out there are some performance improvements we can make with a little rework. More work for less work :)